refactor(types): relocate types to services and update import paths
