+++ app/services/index.ts
@@ -1,2 +1,3 @@
 export * from "./product.server";
 export * from "./productList.server";
+export * from "./types";
